HTML Changes Changed section gallery to slideshow to replace the static images.
Added mulitiple slide elements to hold the images and their descriptions

Added in Next and previous arrows for users to navigate better.

CSS Changes
I styled the slideshow container to ensure consistent spacing with the rest of the page

Added in a fade animation for visual effect when the user changes images

Designed the next and previous buttons so they can  still be seen.

JavaScript Changes
This is the first javascript change to the site

Created the showSlides function to switch between images

added the changeSlide function for the buttons to work

made use to DOMContentLoaded to auto display the first slide when the page is loaded