HTML changes Changed the text in video section
CSS changes

Added in some limits to prevent the nokia image from appearing to large by adding in a max height.

Added in object fit cover to make sure that the images that do get cropped still look okay.

Also added in the styling for the video section for the youtube short.