Started development of snakegame.html
Bulked it out and added what I needed so it worked with the wireframe design.

I created a stylesheet for it as working in the first style sheet was starting to get messy. I have kept consistency with the theme on the previous page also

Created the snake.js file